+++ bolt/passwords.go
@@ -20,6 +20,13 @@ var (
 		Msg:  "your username or password is incorrect",
 	}
 
+	// EIncorrectUser is returned when any user is failed to be found which indicates
+	// the userID provided is for a user that does not exist.
+	EIncorrectUser = &platform.Error{
+		Code: platform.EForbidden,
+		Msg:  "your userID is incorrect",
+	}
+
 	// EShortPassword is used when a password is less than the minimum
 	// acceptable password length.
 	EShortPassword = &platform.Error{
@@ -41,16 +48,16 @@ func CorruptUserIDError(name string, err error) error {
 var _ platform.PasswordsService = (*Client)(nil)
 
 // SetPassword stores the password hash associated with a user.
-func (c *Client) SetPassword(ctx context.Context, name string, password string) error {
+func (c *Client) SetPassword(ctx context.Context, userID platform.ID, password string) error {
 	return c.db.Update(func(tx *bolt.Tx) error {
-		return c.setPassword(ctx, tx, name, password)
+		return c.setPassword(ctx, tx, userID, password)
 	})
 }
 
 // HashCost currently using the default cost of bcrypt
 var HashCost = bcrypt.DefaultCost
 
-func (c *Client) setPassword(ctx context.Context, tx *bolt.Tx, name string, password string) error {
+func (c *Client) setPassword(ctx context.Context, tx *bolt.Tx, userID platform.ID, password string) error {
 	if len(password) < MinPasswordLength {
 		return EShortPassword
 	}
@@ -60,34 +67,33 @@ func (c *Client) setPassword(ctx context.Context, tx *bolt.Tx, name string, pass
 		return err
 	}
 
-	u, pe := c.findUserByName(ctx, tx, name)
-	if pe != nil {
-		return EIncorrectPassword
+	encodedID, err := userID.Encode()
+	if err != nil {
+		return CorruptUserIDError(userID.String(), err)
 	}
 
-	encodedID, err := u.ID.Encode()
-	if err != nil {
-		return CorruptUserIDError(name, err)
+	if _, err := c.findUserByID(ctx, tx, userID); err != nil {
+		return EIncorrectUser
 	}
 
 	return tx.Bucket(userpasswordBucket).Put(encodedID, hash)
 }
 
 // ComparePassword compares a provided password with the stored password hash.
-func (c *Client) ComparePassword(ctx context.Context, name string, password string) error {
+func (c *Client) ComparePassword(ctx context.Context, userID platform.ID, password string) error {
 	return c.db.View(func(tx *bolt.Tx) error {
-		return c.comparePassword(ctx, tx, name, password)
+		return c.comparePassword(ctx, tx, userID, password)
 	})
 }
-func (c *Client) comparePassword(ctx context.Context, tx *bolt.Tx, name string, password string) error {
-	u, pe := c.findUserByName(ctx, tx, name)
-	if pe != nil {
-		return pe
-	}
 
-	encodedID, err := u.ID.Encode()
+func (c *Client) comparePassword(ctx context.Context, tx *bolt.Tx, userID platform.ID, password string) error {
+	encodedID, err := userID.Encode()
 	if err != nil {
 		return err
+	}
+
+	if _, err := c.findUserByID(ctx, tx, userID); err != nil {
+		return EIncorrectUser
 	}
 
 	hash := tx.Bucket(userpasswordBucket).Get(encodedID)
@@ -100,11 +106,11 @@ func (c *Client) comparePassword(ctx context.Context, tx *bolt.Tx, name string, 
 }
 
 // CompareAndSetPassword replaces the old password with the new password if thee old password is correct.
-func (c *Client) CompareAndSetPassword(ctx context.Context, name string, old string, new string) error {
+func (c *Client) CompareAndSetPassword(ctx context.Context, userID platform.ID, old string, new string) error {
 	return c.db.Update(func(tx *bolt.Tx) error {
-		if err := c.comparePassword(ctx, tx, name, old); err != nil {
+		if err := c.comparePassword(ctx, tx, userID, old); err != nil {
 			return err
 		}
-		return c.setPassword(ctx, tx, name, new)
+		return c.setPassword(ctx, tx, userID, new)
 	})
 }